:art: Code restructuring
