--- packages/backend/src/controllers/moodle/config.interface.ts
@@ -1,6 +1,0 @@
-export interface IConfig {
-    plugin: string;
-    subtype: string;
-    name: string;
-    value: string;
-}
