--- packages/backend/src/controllers/moodle/contentfile.interface.ts
@@ -1,9 +1,0 @@
-export interface IContentfile {
-    filename: string;
-    filepath: string;
-    filesize: number;
-    fileurl: string;
-    timemodified: number;
-    mimetype: string;
-    isexternalfile: boolean;
-  }
