--- packages/backend/src/controllers/moodle/course.interface.ts
@@ -1,9 +1,0 @@
-import { IAssignment } from './assignment.interface';
-
-export interface ICourse {
-    id: number;
-    fullname: string;
-    shortname: string;
-    timemodified: number;
-    assignments: IAssignment[];
-}
