--- packages/backend/src/controllers/moodle/ressource.interface.ts
@@ -1,24 +1,0 @@
-import { IContentfile } from './contentfile.interface';
-
-export interface IRessource {
-    id: number;
-    coursemodule: number;
-    course: number;
-    name: string;
-    intro: string;
-    introformat: number;
-    introfiles: any[];
-    contentfiles: IContentfile[];
-    tobemigrated: number;
-    legacyfiles: number;
-    legacyfileslast: any;
-    display: number;
-    displayoptions: string;
-    filterfiles: number;
-    revision: number;
-    timemodified: number;
-    section: number;
-    visible: number;
-    groupmode: number;
-    groupingig: number;
-  }
